+++ example/integration_test/scenario_device_id_init/1.1.BV_dev_type_test.dart
@@ -3,15 +3,15 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import '../utils.dart';
 
-/// 2.Check if the SDK generates a device ID and the type is SDK_GENERATED
+/// 1.1.Test Device ID Type - Bad Value - Developer sets the ID during init
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-  testWidgets('2.Test Device ID Type - SDK generates ID', (WidgetTester tester) async {
+  testWidgets('1.1.Test Device ID Type - Bad Value - Developer sets the ID during init', (WidgetTester tester) async {
     // Initialize the SDK
-    CountlyConfig config = CountlyConfig(SERVER_URL, APP_KEY).setLoggingEnabled(true);
+    CountlyConfig config = CountlyConfig(SERVER_URL, APP_KEY).setLoggingEnabled(true).setDeviceId('');
     await Countly.initWithConfig(config);
 
-    final id = await testDeviceID(isNot('test'));
+    final id = await testDeviceID(isNot(''));
     await testDeviceIDType(DeviceIdType.SDK_GENERATED);
     await testLastRequestParams({'device_id': id});
   });